bugfixes+2nd auton
fixed an incorrectly assigned controller axis, resulting in the robot moving when turning the joystick left or right rather than up and down. Mirrored an auton program to give us options for being on red instead of blue. Set the preauton to trigger the driver() function for easy testing. All the vex ports seem broken tho on our bot :(