Auton work
Created auton for the other part of the field and mirrored it to the other side. Fixed miscalculations in distance in earlier auton program. updated preauton to include all 4. Still need to create auton skills. Main is still adjusted for testing driver and not auton. Need to adjust main to work better with auton